Fixes for usage in combination with loadwatch
This fixes the problems described in bug #230.

When using jline to create a new ConsoleReader object
in combination with loadwatch, the process never returned.

In jline1, a long time ago, Runtime.getRuntime().exec(cmd)
was used to execute commands. This was migrated to the
ProcessBuilder API later, which is the recommended way.

In the meantime, the commands (as plain strings) were split
manually with split(""), whereas Runtime.getRuntime().exec(cmd)
was using a StringTokenizer internally.

The java implementations here are different, and one needs to
be carefull when passing down stuff to the ProcessBuilder.
See also:
http://stackoverflow.com/questions/6856028/difference-between-processbuilder-and-runtime-exec

My implementation just uses explicit List<String> calls everywhere
to make clear how arguments are split, which fixes the usage of
jline in combination with loadwatch.