Proposed fix for #272
`SecurityException` can be thrown when converting `File` to `URL`
and the code was not taking this into account.

This fix accepts that it will propagate up to the point where
the file is read, and the `SecurityException` is caught there,
since opening the file could throw one as well anyway.